modified paystack id field in transaction model to big int

+++ _base/payments/models.py
@@ -32,7 +32,7 @@ class Transaction(BaseModel):
         default=None
     )
 
-    paystack_id = models.IntegerField(
+    paystack_id = models.BigIntegerField(
         null=True,
         default=None,
     )
